revamp search.css for articles
TODO: fix rest of css for article. I've only done the button and the placement for the text on the left. The tags should be spaced out on a line to the left and right and the right container should be extended. The text should get the correct colors and spacings.